Disallow `module {` at the start of expr statemets or default exports
For forward compatibility with the [module declarations](https://github.com/tc39/proposal-module-fragments)
proposal, we need to disallow `export default module {}`, because it
will be a module declaration and not a module expression.

After module declartions, for consistency with functions and classes
we may have named module expressions: in that case `module Identifier {`
will also need to be disallowed from starting an expression statement
because it must be parsed as a module declaration.

`module {` would still be a valid expression statement start, but we
disallow it in this PR for consistency with:
- possible future named module expressions
- function&classes, that are disallowed even when they don't have a name
  and thus would be unambiguously expression statements.